% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSMDM.R
\name{fitSMDM}
\alias{fitSMDM}
\title{Fit MM-type regression}
\usage{
fitSMDM(formula, data, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit. See 
\code{\link[stats]{lm}} and \code{\link[stats]{formula}} for more details.}

\item{data}{an optional data frame, list or environment 
(or object coercible by \code{\link[base]{as.data.frame}} to a data frame) 
containing the variables in the model. If not found in data, 
the variables are taken from environment(formula), 
typically the environment from which lmrob is called.}

\item{...}{additional arguments can be used to specify control parameters directly instead of (but not in addition to!) via control.}
}
\value{
An object of class \code{\link[robustbase]{lmrob}}
}
\description{
Fit MM-type regression with \code{\link[robustbase]{lmrob}}
}
\details{
see \code{\link[robustbase]{lmrob}}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[robustbase]{lmrob}}
}
