% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_eq.R
\name{lm_eq}
\alias{lm_eq}
\title{Expression linear model equation and r-square.}
\usage{
lm_eq(df, method)
}
\arguments{
\item{df}{input dataframe including columns "Sample.VWC" \code{y} and "Sensor.VWC"  \code{x}}

\item{method}{define method, \code{lm} for OLS estimator, \code{rlm} for MM-type estimator (robust).}
}
\value{
character of linear model equation and R-square value
}
\description{
Computes and returns (as.character) linear model equation and R-square value.
}
\details{
DETAILS
}
\examples{
data("data")
lm_eq(df = data, method = "rlm")
}
