% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyapp.R
\name{runShinyapp}
\alias{runShinyapp}
\title{run Shiny app for soil moisture sensor calibration}
\usage{
runShinyapp()
}
\value{
shiny gui
}
\description{
\code{runShinyapp} does start shiny application for soil moisture sensor calibration.
}
\examples{
SWCalibrateR::runShinyapp()
}
\seealso{
\code{\link[shiny]{runApp}}
}
