% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{server for shiny app}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{input data set}

\item{output}{output list}

\item{session}{session info}
}
\description{
server for soil moisture calibration shiny app
}
\examples{
runApp(ui, server)
}
\seealso{
\code{\link[dplyr]{filter}}
 \code{\link[ggplot2]{ggplot}},\code{\link[ggplot2]{aes}},\code{\link[ggplot2]{geom_abline}},\code{\link[ggplot2]{geom_point}},\code{\link[ggplot2]{coord_cartesian}},\code{\link[ggplot2]{facet_grid}},\code{\link[ggplot2]{geom_label}},\code{\link[ggplot2]{geom_smooth}}
 \code{\link[leaflet]{leafletOutput}},\code{\link[leaflet]{awesomeIcons}},\code{\link[leaflet]{leaflet}},\code{\link[leaflet]{addProviderTiles}},\code{\link[leaflet]{addAwesomeMarkers}},\code{\link[leaflet]{addMeasure}},\code{\link[leaflet]{addLayersControl}}
 \code{\link[leaflet.extras]{addSearchOSM}}
}
