% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMCcalibration.R
\docType{package}
\name{SMCcalibration}
\alias{SMCcalibration}
\alias{SMCcalibration-package}
\title{SMCcalibration: A package for estimate and diagnose soil moisture sensor
calibration.}
\description{
The SMCcalibration package provides an interactive shiny framework.
}
\section{SMCcalibration functions}{

The SMCcalibration functions ...···
}

