% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWCalibrateR.R
\docType{package}
\name{SWCalibrateR}
\alias{SWCalibrateR}
\alias{SWCalibrateR-package}
\title{SWCalibrateR: A package for estimate and diagnose calibration equations for soil moisture sensors}
\description{
The SWCalibrateR package provides an interactive shiny framework.
}
